examples/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []  # This is a project directory, not a package

[tool.setuptools.package-data]
"*" = []

[project]
name = "interlace-api-examples"
version = "0.1.0"
description = "Example Interlace project consuming REST APIs"
requires-python = ">=3.12"
dependencies = [
    "interlace",  # Referenced from workspace member (pkg)
    "ibis-framework[duckdb]>=9.0.0",
    "pandas>=2.0.0",
    "aiohttp>=3.9.0",  # For API helper
]

[tool.uv.sources]
interlace = { workspace = true }

